backend/requirements.txt
# Backend-specific requirements
# Core application dependencies are in root requirements.txt

# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic[email]>=2.5.0

# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# AI/ML
openai>=1.3.0
anthropic>=0.7.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Email processing
beautifulsoup4>=4.12.0
html2text>=2020.1.16

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

